# giggz-api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /auth/validateToken:
    get:
      summary: Validate JWT token
      description: Validates the provided JWT token to check if it's valid and not expired.
      operationId: validateToken
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token to be validated.
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        200:
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token is valid"
        401:
          description: Invalid or expired token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid token"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT